ai/src/requirements.txt
pymongo==4.8.0
python-dotenv==1.0.1
dnspython==2.6.1
python-pptx==0.6.23
Pillow==10.4.0